requirements.txt
Flask==0.9
Flask-Login==0.1.3
Flask-Mail==0.7.4
Flask-OAuth==0.12
Flask-Principal==0.3.3
Flask-Security==1.6.2
Flask-Social==1.5.5
Flask-Uploads==0.1.3
Flask-WTF==0.8.2
Jinja2==2.6
Pillow==2.0.0
PyYAML==3.10
WTForms==1.0.3
Werkzeug==0.8.3
blinker==1.2
flask-mongoengine==0.6
httplib2==0.8
itsdangerous==0.17
mongoengine==0.7.9
oauth2==1.5.211
passlib==1.6.1
pycrypto==2.6
pymongo==2.4.2
requests==1.2.0
stripe==1.7.10
wsgiref==0.1.2
